Add flash message to order controller create action

+++ app/controllers/orders_controller.rb
@@ -13,23 +13,29 @@ class OrdersController < ApplicationController
   end
 
   def create
-    order = current_user.orders.create(order_params)
-
-    if order.save
-      cart.items.each do |item,quantity|
-        order.item_orders.create({
-          item: item,
-          quantity: quantity,
-          price: item.price
-          })
-      end
-      session.delete(:cart)
-      flash[:success] = "Your order has been created!"
-      # redirect_to "/orders/#{order.id}"
-       redirect_to "/profile/orders"
-    else
-      flash[:notice] = "Please complete address form to create an order."
+    if !current_user
+      flash[:warning] = "Must log in before checking out"
       render :new
+    else
+
+      order = current_user.orders.create(order_params)
+
+      if order.save
+        cart.items.each do |item,quantity|
+          order.item_orders.create({
+            item: item,
+            quantity: quantity,
+            price: item.price
+            })
+        end
+        session.delete(:cart)
+        flash[:success] = "Your order has been created!"
+        # redirect_to "/orders/#{order.id}"
+         redirect_to "/profile/orders"
+      else
+        flash[:notice] = "Please complete address form to create an order."
+        render :new
+      end
     end
   end
